GCP/F2/requirement.txt
google-cloud-storage==1.42.0
google-auth==2.0.2
google-auth-oauthlib==0.4.5
google-auth-httplib2==0.1.0
